pyproject.toml
[project]
name = "cpast_llm"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]

[tool.poetry]
name = "cpast-llm"
version = "0.1.0"
description = "LLM for cpast lib written in Rust to generate clex language from human readable input format and constarints"
authors = ["Lab Rat <dev.frolics@gmail.com>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain-google-genai = "^0.0.6"
python-dotenv = "^1.0.1"
langchain = "^0.1.4"
fastapi = "^0.109.0"
uvicorn = "^0.27.0.post1"
requests = "^2.31.0"
requests-cache = "^1.1.1"
pylatexenc = "^2.10"
bs4 = "^0.0.2"
html5lib = "^1.1"

[tool.poetry.group.dev.dependencies]
maturin = "^1.4.0"
pytest = "^8.0.0"
ruff = "^0.2.1"
pyright = "^1.1.350"

[build-system]
requires = ["maturin>=1.4,<2.0"]
build-backend = "maturin"

[tool.maturin]
features = ["pyo3/extension-module"]

[tool.ruff.lint]
# 1. Enable flake8-bugbear (`B`) rules, in addition to the defaults.
select = ["E4", "E7", "E9", "F", "B"]

# 2. Avoid enforcing line-length violations (`E501`)
ignore = ["E501"]

# 3. Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]

# 4. Ignore `E402` (import violations) in all `__init__.py` files, and in select subdirectories.
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402"]

[tool.ruff.format]
# 5. Use single quotes for non-triple-quoted strings.
quote-style = "single"

[tool.pyright]
include = ["cpast_llm", "cpast_scrapper", "main.py", "app.py", "tests", "scrape.py"]
exclude = ["**/__pycache__",
    "**/__init__.py"
]
defineConstant = { DEBUG = true }

reportMissingImports = true
reportMissingTypeStubs = false

pythonVersion = "3.10"
pythonPlatform = "Linux"
